Fix UI bugs
