--- src/main/resources/view/KanbanBoard.fxml
@@ -1,98 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import java.net.URL?>
-<?import javafx.geometry.Insets?>
-<?import javafx.scene.Scene?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.Menu?>
-<?import javafx.scene.control.MenuBar?>
-<?import javafx.scene.control.MenuItem?>
-<?import javafx.scene.layout.HBox?>
-<?import javafx.scene.layout.StackPane?>
-<?import javafx.scene.layout.VBox?>
-<?import javafx.scene.shape.Line?>
-<?import javafx.stage.Stage?>
-
-<fx:root minHeight="600" minWidth="450" onCloseRequest="#handleExit" title="Kanbug Tracker" type="javafx.stage.Stage" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1">
-    <scene>
-        <Scene>
-            <stylesheets>
-                <URL value="@DarkTheme.css" />
-                <URL value="@Extensions.css" />
-            </stylesheets>
-            <VBox>
-            <children>
-                <MenuBar fx:id="menuBar" VBox.vgrow="NEVER">
-                    <Menu mnemonicParsing="false" text="File">
-                        <MenuItem mnemonicParsing="false" onAction="#handleExit" text="Exit" />
-                    </Menu>
-                    <Menu mnemonicParsing="false" text="Help">
-                        <MenuItem fx:id="helpMenuItem" mnemonicParsing="false" onAction="#handleHelp" text="Help" />
-                    </Menu>
-                </MenuBar>
-                <StackPane fx:id="commandBoxPlaceholder" styleClass="pane-with-border" VBox.vgrow="NEVER">
-                    <padding>
-                        <Insets bottom="5" left="10" right="10" top="5" />
-                    </padding>
-                </StackPane>
-                <StackPane fx:id="resultDisplayPlaceholder" maxHeight="100" minHeight="100" prefHeight="100" styleClass="pane-with-border" VBox.vgrow="NEVER">
-                    <padding>
-                        <Insets bottom="5" left="10" right="10" top="5" />
-                    </padding>
-                </StackPane>
-            <HBox fx:id="board" alignment="CENTER" HBox.hgrow="ALWAYS">
-                <children>
-                    <VBox alignment="CENTER" styleClass="pane-with-border" HBox.hgrow="ALWAYS">
-                        <children>
-                            <Label fx:id="backlogLabel" contentDisplay="CENTER" text="Backlog" textAlignment="CENTER" styleClass="state" HBox.hgrow="ALWAYS" VBox.vgrow="ALWAYS">
-                                <HBox.margin>
-                                    <Insets right="5.0" />
-                                </HBox.margin>
-                            </Label>
-                            <StackPane fx:id="bugListPanelPlaceholderBacklog" HBox.Hgrow="ALWAYS" />
-                        </children>
-                    </VBox>
-                    <Line/>
-                    <VBox alignment="CENTER" styleClass="pane-with-border" HBox.hgrow="ALWAYS">
-                        <children>
-                            <Label fx:id="todoLabel" alignment="CENTER" contentDisplay="CENTER" styleClass="state" text="To Do" textAlignment="CENTER" VBox.vgrow="ALWAYS">
-                                <HBox.margin>
-                                    <Insets right="5.0" />
-                                </HBox.margin>
-                            </Label>
-                            <StackPane fx:id="bugListPanelPlaceholderTodo" HBox.Hgrow="ALWAYS" />
-                        </children>
-                    </VBox>
-                    <Line />
-                    <VBox alignment="CENTER" styleClass="pane-with-border" HBox.hgrow="ALWAYS">
-                        <children>
-                            <Label fx:id="ongoingLabel" alignment="CENTER" contentDisplay="CENTER" styleClass="state" text="Ongoing" textAlignment="CENTER" VBox.vgrow="ALWAYS">
-                                <HBox.margin>
-                                    <Insets right="5.0" />
-                                </HBox.margin>
-                            </Label>
-                            <StackPane fx:id="bugListPanelPlaceholderOngoing" HBox.Hgrow="ALWAYS" />
-                        </children>
-                    </VBox>
-                    <Line />
-                    <VBox alignment="CENTER" styleClass="pane-with-border" HBox.hgrow="ALWAYS">
-                        <children>
-                            <Label fx:id="doneLabel" alignment="CENTER" contentDisplay="CENTER" styleClass="state" text="Done" textAlignment="RIGHT" VBox.vgrow="ALWAYS">
-                                <HBox.margin>
-                                    <Insets right="5.0" />
-                                </HBox.margin>
-                              <opaqueInsets>
-                                 <Insets left="5.0" />
-                              </opaqueInsets>
-                            </Label>
-                            <StackPane fx:id="bugListPanelPlaceholderDone" HBox.Hgrow="ALWAYS" />
-                        </children>
-                    </VBox>
-                </children>
-            </HBox>
-                <StackPane fx:id="statusbarPlaceholder" VBox.vgrow="NEVER" />
-            </children>
-            </VBox>
-        </Scene>
-    </scene>
-</fx:root>
