+++ src/main/resources/view/MainWindow.fxml
@@ -51,6 +51,7 @@
             <Insets top="10" right="10" bottom="10" left="10" />
           </padding>
           <StackPane fx:id="bugListPanelPlaceholder" VBox.vgrow="ALWAYS"/>
+          <StackPane fx:id="fourBugListPanelsPlaceholder" />
         </VBox>
 
         <StackPane fx:id="statusbarPlaceholder" VBox.vgrow="NEVER" />
